Updated repository to be opened from eclipse
The folder "MyRobots" is the java project folder, which will contain the
source code for each robot.  All modified source should be from the
folder "/MyRobots/src" and not the "/src" folder.

Additionally, you can simply open up the "robot" folder as an Eclipse
workspace, instead of copying/pasting source files to and from the
eclipse/git folder.