Fixed movement and radar, added ramming.
We had to change to an advanced bot from a rate control robot for now,
in order to use the Goto method.  We may want to change this in the
future, but rates are much more annoying than goto.

The radar constantly spins for now, adding and updating other enemies to
the HashMap named "enemies".  This map will contain everything we can
get from the onRobotScanned event.  The key of the map is the enemy
robot string name.

Finally, I added ramming to the robot.  If we are close to an enemy,
then the robot will make the enemies last location its new goto
location.  Note that goto is called on every tick, so it is
significantly better than ramfire.